Add files via upload
It looks like most values are on the low end of each attribute and low corresponding cancer rates.  When attribute values are on the high end, there is a higher likelihood of cancer, especially A3 (Uniformity of Cell Size), A4 (Uniformity of Cell Shape), A7 (Bare Nuclei), which have the higher correlation values.